+++ D3D11Engine/GothicAPI.h
@@ -11,7 +11,6 @@
 #include "RenderQueue.h"
 #include "RenderToTextureBuffer.h"
 #include "ShaderIDs.h"
-#include "BspPortalCuller.h"
 
 static const char* MENU_SETTINGS_FILE = "system\\GD3D11\\UserSettings.ini";
 const float INDOOR_LIGHT_DISTANCE_SCALE_FACTOR = 0.5f;
@@ -39,7 +38,6 @@ struct RndCullContext {
     RenderStage stage;
 
     RenderQueue* queue;
-    const class BspPortalCuller* portalCuller = nullptr;
     
     struct
     {
@@ -100,7 +98,6 @@ struct BspInfo {
         IndoorLights = std::move( other.IndoorLights );
         Mobs = std::move( other.Mobs );
         NodePolygons = std::move( other.NodePolygons );
-        SectorIds = std::move( other.SectorIds );
         NumStaticLights = other.NumStaticLights;
         OriginalNode = other.OriginalNode;
         Front = other.Front;
@@ -131,10 +128,6 @@ struct BspInfo {
     zCBspBase* OriginalNode;
     BspInfo* Front;
     BspInfo* Back;
-
-    /** Sectors (rooms) this leaf holds polys of, as indices into BspPortalCuller's sector array.
-        Empty on outdoor leafs, which is the vast majority - see BspPortalCuller::BuildFromWorld. */
-    std::vector<uint16_t> SectorIds;
 };
 
 /** Pre-built linear cache of all BSP leaf bounding boxes for SIMD-accelerated frustum culling.
@@ -256,8 +249,6 @@ class zCDecal;
 class GothicAPI {
     friend struct MaterialInfo;
 public:
-    BspPortalCuller& GetPortalCuller() { return PortalCuller; }
-    const BspPortalCuller& GetPortalCuller() const { return PortalCuller; }
     GothicAPI();
     ~GothicAPI();
 
@@ -848,7 +839,6 @@ private:
     std::vector<SkeletalVobInfo*> AnimatedSkeletalVobs;
     std::vector<TransparencyVobInfo> TransparencyVobs;
     std::vector<SkeletalVobInfo*> VNSkeletalVobs;
-    gtl::flat_hash_map<zCVob*, SkeletalVobInfo*> SkeletalVobMap;
 
     /** List of Vobs having a zCParticleFX-Visual */
     std::vector<zCVob*> ParticleEffectVobs;
@@ -895,7 +885,7 @@ public:
     // Exposed for CollectLeafVobs/CollectVisibleVobsWithLeafCache (file-static helpers)
     BspLeafLinearCache LeafLinearCache;
 private:
-    BspPortalCuller PortalCuller;
+    gtl::flat_hash_map<zCVob*, SkeletalVobInfo*> SkeletalVobMap;
 
     /** Map of VobInfo-Lists for zCBspLeafs */
     std::unordered_map<zCBspBase*, BspInfo> BspLeafVobLists;