+++ D3D11Engine/zCMaterial.h
@@ -10,7 +10,6 @@
 const int zMAT_GROUP_WATER = 5;
 const int zMAT_GROUP_SNOW = 6;
 
-class zCBspSector;
 class zCTexAniCtrl {
 private:
     int	AniChannel;
@@ -173,12 +172,6 @@ public:
     zTRnd_AlphaBlendFunc GetAlphaFunc() {
         MaterialFlags& flags = GetFlags();
         return flags.rndAlphaBlendFunc;
-    }
-    zCBspSector* GetBspSectorFront() const {
-        return *reinterpret_cast<zCBspSector* const*>(THISPTR_OFFSET( GothicMemoryLocations::zCMaterial::Offset_BspSectorFront ));
-    }
-    zCBspSector* GetBspSectorBack() const {
-        return *reinterpret_cast<zCBspSector* const*>(THISPTR_OFFSET( GothicMemoryLocations::zCMaterial::Offset_BspSectorBack ));
     }
 
     void SetAlphaFunc( zTRnd_AlphaBlendFunc func ) {